refactor: pages and src/pages folders
