--- pages/auth/createnewpassword/index.tsx
@@ -1,1 +1,0 @@
-export { CreateNewPasswordPage as default } from '@/pages/createNewPassword'
