--- pages/auth/forgotpassword/index.tsx
@@ -1,1 +1,0 @@
-export { ForgotPasswordPage as default } from '@/pages/forgotPassword'
