+++ pages/index.tsx
@@ -1,2 +1,1 @@
-import Home from './home'
-export default Home
+export { Home as default } from '@/pages/home'
